requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-community
chromadb
sentence-transformers
python-dotenv
pydantic
langchain-huggingface
langchain-chroma
langchain-core